rag-query-ms/pyproject.toml
[project]
name = "rag-query-ms"
version = "0.1.0"
description = "A microservice for querying a RAG pipeline."
requires-python = ">=3.13"
dependencies = [
    "fastapi",
    "uvicorn[standard]",
    "pydantic-settings",
    "langchain",
    "langchain-community",
    "chromadb-client",
    "langchain-huggingface>=0.3.1",
    "langchain-chroma>=0.2.5",
    "langchain-google-genai",
    "langgraph",      
    "loguru",
    "gunicorn"
]

[project.optional-dependencies]
dev = [
    "pytest",
    "httpx" # Asynchroniczny klient HTTP do testowania endpointów FastAPI
]

[tool.uv.sources]
pypi = { url = "https://pypi.org/simple" }